abstract query code for esier implementation
